% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe.R
\name{npe}
\alias{npe}
\title{Compute the Neighbourhood Proportion Error}
\usage{
npe(
  hd,
  ld,
  annot,
  knn = NULL,
  k = NULL,
  normalise = FALSE,
  exclude_pops = c(),
  metric = "total_variation_distance",
  reduce = "sum",
  plot_path = NULL
)
}
\arguments{
\item{hd}{numeric matrix: matrix of row-wise coordinate vectors of points in high-dimensinal (HD, original) space}

\item{ld}{numeric matrix: matrix of row-wise coordinate vectors of points in lower-dimensional (LD, reduced) space}

\item{annot}{factor or character vector: label per each row of \code{hd} and \code{ld}}

\item{knn}{\emph{k}-nearest-neighbour indices in the original space (in the same format as output of \code{FNN::knn}. Default value is \code{NULL}; this triggers construction of a new \emph{k}-NN graph}

\item{k}{integer: neighbourhood size for \emph{k}-NN graph construction (if \code{knn} is not \code{NULL}). Default value is \code{NULL}}

\item{normalise}{logical: whether \code{hd} and \code{ld} should be z-scaled column-wise prior to \emph{k}-NN graph computation. Default value is \code{FALSE}}

\item{metric}{character: name of metric to use for measuring dissimilarity between likeness distributions for each labelled population. One of \emph{'total_variation_distance'} (or \emph{'tvd'}) and \emph{'earth_movers_distance'} (or \emph{'emd'}). Default value is \emph{'total_variation_distance'}}

\item{reduce}{character: name of reduction method to convert the vector of dissimilarity values between distributions to a single numeric value. One of \emph{'sum'}, \emph{'mean'} (or \emph{'average'}, \emph{'avg'}) and \emph{'none'} (to get the original vector of values). Default value is \emph{'sum'}}

\item{plot_path}{character: if specified, this is a path for saving a PDF containing diagram with the distributions for each population in HD and LD. Default value is \code{NULL} (no PDF is produced)}
}
\description{
Computes the Neighbourhood Proportion Error (NPE), a quality-assessment measure for structure preservation by lower-dimensional embeddings of labelled high-dimensional data.
}
\details{
The original NPE manuscript: Konstorum A, Jekel N, Vidal E, Laubenbacher R (2018) Comparative Analysis of Linear and Nonlinear Dimension Reduction Techniques on Mass Cytometry Data, bioRxiv 273862; doi: https://doi.org/10.1101/273862.
Original GitHub repo: https://github.com/akonstodata/cytof_dimred.
Modified version written by David Novak.
}
